feat: split Icon into icons and component
