+++ index.js
@@ -1,0 +1,3 @@
+import icons from './src/icons'
+
+export icons
