+++ loaders/strip-fill-loader.js
